NN-4170 refactoring per PR comments, add feature flag
