+++ feature.env
@@ -13,3 +13,4 @@ CURIOUS_API_URL=http://localhost:9091/curious
 TOKEN_VERIFICATION_ENABLED=true
 GOTENBERG_API_URL=http://localhost:3001
 ADJUDICATIONS_URL=http://host.docker.internal:3007
+YOI_NEW_PAGES_FEATURE_FLAG=true
