+++ integration_tests/mockApis/adjudications.ts
@@ -140,6 +140,29 @@ const stubEditDraftIncidentDetails = ({
     },
   })
 
+const stubUpdateDraftIncidentRole = ({
+  id,
+  response = {},
+  status = 200,
+}: {
+  id: number
+  response: Record<string, unknown>
+  status: number
+}): SuperAgentRequest =>
+  stubFor({
+    request: {
+      method: 'PUT',
+      url: `/adjudications/draft-adjudications/${id}/incident-role`,
+    },
+    response: {
+      status,
+      headers: {
+        'Content-Type': 'application/json;charset=UTF-8',
+      },
+      jsonBody: response,
+    },
+  })
+
 const stubGetReportedAdjudication = ({
   id,
   response = {},
@@ -329,6 +352,7 @@ export default {
   stubGetDraftAdjudication,
   stubSubmitCompleteDraftAdjudication,
   stubEditDraftIncidentDetails,
+  stubUpdateDraftIncidentRole,
   stubGetReportedAdjudication,
   stubGetAllDraftAdjudicationsForUser,
   stubGetYourReportedAdjudications,
