+++ server/config.ts
@@ -116,4 +116,5 @@ export default {
   domain: get('INGRESS_URL', 'http://localhost:3000', requiredInProduction),
   digitalPrisonServiceUrl: get('DIGITAL_PRISON_SERVICE_URL', 'http://localhost:3002', requiredInProduction),
   supportUrl: get('SUPPORT_URL', 'http://localhost:3003', requiredInProduction),
+  yoiNewPagesFeatureFlag: get('YOI_NEW_PAGES_FEATURE_FLAG', true, requiredInProduction),
 }
