--- server/routes/incidentRole/incidentRoleEdit.ts
@@ -1,19 +1,0 @@
-import { Request, Response } from 'express'
-import PlaceOnReportService from '../../services/placeOnReportService'
-import IncidentRolePage, { PageRequestType } from './incidentRolePage'
-
-export default class IncidentRoleEditRoutes {
-  page: IncidentRolePage
-
-  constructor(placeOnReportService: PlaceOnReportService) {
-    this.page = new IncidentRolePage(PageRequestType.EDIT, placeOnReportService)
-  }
-
-  view = async (req: Request, res: Response): Promise<void> => {
-    await this.page.view(req, res)
-  }
-
-  submit = async (req: Request, res: Response): Promise<void> => {
-    await this.page.submit(req, res)
-  }
-}
