+++ server/routes/index.ts
@@ -24,6 +24,7 @@ import deletePersonRoutes from './deletePerson'
 import { Services } from '../services'
 import adjudicationPdfRoutes from './adjudicationPdf'
 import adjudicationUrls from '../utils/urlGenerator'
+import config from '../config'
 
 export default function routes(
   router: Router,
@@ -43,7 +44,9 @@ export default function routes(
     offenceCodeDecisionsRoutes({ placeOnReportService, userService, offenceSessionService, decisionTreeService })
   )
   router.use(adjudicationUrls.incidentDetails.root, incidentDetailsRoutes({ placeOnReportService, locationService }))
-  router.use(adjudicationUrls.incidentRole.root, incidentRoleRoutes({ placeOnReportService }))
+  if (config.yoiNewPagesFeatureFlag) {
+    router.use(adjudicationUrls.incidentRole.root, incidentRoleRoutes({ placeOnReportService }))
+  }
   router.use(
     adjudicationUrls.detailsOfOffence.root,
     detailsOfOffenceRoutes({ placeOnReportService, allOffencesSessionService, decisionTreeService })
