+++ server/routes/util/incidentSearchValidation.ts
